Ensure Erlang VM is stopped when RabbitMQ stops
[#128761867]

+++ jobs/rabbitmq-server/templates/drain.bash
@@ -11,16 +11,50 @@ SHUTDOWN_LOG="${LOG_DIR}"/shutdown_stdout.log
 SHUTDOWN_ERR_LOG="${LOG_DIR}"/shutdown_stderr.log
 
 main() {
-  log "Attempting to stop RabbitMQ instance..."
-  rabbitmqctl stop "$ERLANG_PID_FILE" 1>> "$SHUTDOWN_LOG" 2>> "$SHUTDOWN_ERR_LOG"
-  # ensure RabbitMQ is stopped
-  rm -f "$ERLANG_PID_FILE"
-  log "RabbitMQ instance stopped successfully"
+  log "Begin RabbitMQ node shutdown ..."
+
+  if rabbitmq_node_is_stopped
+  then
+    log "RabbitMQ node is not running, nothing to shutdown."
+  else
+    rabbitmq_node_is_healthy
+    cluster_is_healthy
+    log "Stop RabbitMQ node..."
+    rabbitmqctl stop "$ERLANG_PID_FILE" 1>> "$SHUTDOWN_LOG" 2>> "$SHUTDOWN_ERR_LOG"
+    log "Checking RabbitMQ node is stopped ..."
+    rabbitmq_node_is_stopped
+    rm -f "$ERLANG_PID_FILE"
+    log "RabbitMQ node stopped successfully."
+  fi
+
   echo "0"
 }
 
 log() {
   echo "$*" 1>> "$SHUTDOWN_LOG"
+}
+
+rabbitmq_node_is_stopped() {
+  if erlang_pid_file_exists
+  then
+    ! erlang_pid_points_to_a_running_process
+  fi
+}
+
+erlang_pid_file_exists() {
+  [ -f "$ERLANG_PID_FILE" ]
+}
+
+erlang_pid_points_to_a_running_process() {
+  ps "$(cat $ERLANG_PID_FILE)" > /dev/null
+}
+
+rabbitmq_node_is_healthy() {
+  log "Check RabbitMQ node is healthy ..."
+}
+
+cluster_is_healthy() {
+  log "Check RabbitMQ cluster is healthy ..."
 }
 
 main
